_Investment_v2/requirements.txt
openai
pydantic
ccxt
pandas
pandas_ta
python-dotenv
termcolor
python-binance
fastapi
uvicorn
python-multipart
websockets
wsproto
motor
tiktoken
